requirements.txt
fastapi
uvicorn
gunicorn
python-multipart>=0.0.6
azure-storage-blob>=12.16.0
azure-search-documents>=11.4.0
openai>=1.6.0
python-dotenv
pydantic==2.4.2
streamlit>=1.22.0
requests>=2.28.2
